ConsoleActivity: fix on-screen button-group auto hiding interfering with clicks
As soon as the on-screen group is made visible, its hiding is scheduled.

In case of a click on one of the buttons the group is instantly hidden.

The problem was, that the scheduled task was not removed in this case.
This could lead to unexpected behaviour if the group was opened again,
while the scheduled task was not yet performed -> the group would disappear
too early.

Solve it by removing the the Runnable from the Handler when group is hidden.