ConsoleActivity: properly handle orientation change for API 13+
added 'screenSize' to 'android:configChanges', because it gets
triggerd additionally to 'orientation' on API 13+.

From the docs:
> Note: If your application targets API level 13 or higher
> (as declared by the minSdkVersion and targetSdkVersion attributes),
> then you should also declare the "screenSize" configuration,
> because it also changes when a device switches between
> portrait and landscape orientations.

This restores the intended behaviour before the bump to API 15,
which speeds up the orientation change a lot, because the whole
activity destory/create part is skipped.

This also fixes a bug where all menu entries were disabled after
an orientation change.