Use proper theme depending on android version and make it consistent in host list and terminal
Before: The terminal session forced Theme.Holo (NoTitle), the rest of the app
used whatever was default on the device. This could lead to different themes
being applied in the host list and in the terminal, if the device is using
a different default theme (android 14+).

Now: Set a new DefaultTheme for the whole application:
Use 'Theme' for android < 11
Use 'Theme.Holo' for android < 14
Use 'Theme.DeviceDefault' for android >= 14
This is essentially the same as not specifying the theme at all.

Then inherit the NoTitle theme for the terminal from that one. Now the theme
is consistent in all parts of the app.

In the future, a theme selector might be added to the options, to override
this behavior, and, for example use the Holo theme on jellybean, even if the
device manufactor bundled a different default theme (galaxy s3 for example).